+++ packages/flutter_riverpod/test/consumer_test.dart
@@ -2,9 +2,193 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_riverpod/src/internal.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/mockito.dart';
 import 'package:state_notifier/state_notifier.dart';
 
 void main() {
+  testWidgets(
+      'Consumer removing one of multiple listeners on a provider still listen to the provider',
+      (tester) async {
+    final stateProvider = StateProvider((ref) => 0, name: 'state');
+    final notifier0 = TestNotifier();
+    final notifier1 = TestNotifier(42);
+    final provider0 = StateNotifierProvider((_) => notifier0, name: '0');
+    final provider1 = StateNotifierProvider((_) => notifier1, name: '1');
+    var buildCount = 0;
+
+    await tester.pumpWidget(ProviderScope(
+      child: Consumer((c, read) {
+        buildCount++;
+        final state = read(stateProvider).state;
+        final value =
+            state == 0 ? read(provider0.state) : read(provider1.state);
+
+        return Text(
+          '${read(provider0.state)} $value',
+          textDirection: TextDirection.ltr,
+        );
+      }),
+    ));
+    final owner = tester //
+        .state<ProviderScopeState>(find.byType(ProviderScope))
+        .owner;
+
+    provider0.state.readOwner(owner);
+    provider1.state.readOwner(owner);
+    final familyState0 = owner.debugProviderStates.firstWhere((p) {
+      return p.provider == provider0.state;
+    });
+    final familyState1 = owner.debugProviderStates.firstWhere((p) {
+      return p.provider == provider1.state;
+    });
+
+    expect(buildCount, 1);
+    expect(familyState0.$hasListeners, true);
+    expect(familyState1.$hasListeners, false);
+    expect(find.text('0 0'), findsOneWidget);
+
+    notifier0.increment();
+    await tester.pump();
+
+    expect(buildCount, 2);
+    expect(find.text('1 1'), findsOneWidget);
+
+    notifier1.increment();
+    await tester.pump();
+
+    expect(buildCount, 2);
+
+    // changing the provider that computed is subscribed to
+    stateProvider.readOwner(owner).state = 1;
+    await tester.pump();
+
+    expect(buildCount, 3);
+    expect(find.text('1 43'), findsOneWidget);
+    expect(familyState1.$hasListeners, true);
+    expect(familyState0.$hasListeners, true);
+
+    notifier1.increment();
+    await tester.pump();
+
+    expect(buildCount, 4);
+    expect(find.text('1 44'), findsOneWidget);
+
+    notifier0.increment();
+    await tester.pump();
+
+    expect(buildCount, 5);
+    expect(find.text('2 44'), findsOneWidget);
+  });
+  testWidgets(
+      'Stops listening to a provider when recomputed but no longer using it',
+      (tester) async {
+    final stateProvider = StateProvider((ref) => 0, name: 'state');
+    final notifier0 = TestNotifier();
+    final notifier1 = TestNotifier(42);
+    final provider0 = StateNotifierProvider((_) => notifier0, name: '0');
+    final provider1 = StateNotifierProvider((_) => notifier1, name: '1');
+    var buildCount = 0;
+
+    await tester.pumpWidget(
+      ProviderScope(
+        child: Consumer((c, read) {
+          buildCount++;
+          final state = read(stateProvider).state;
+          final result = state == 0 //
+              ? read(provider0.state)
+              : read(provider1.state);
+          return Text('$result', textDirection: TextDirection.ltr);
+        }),
+      ),
+    );
+    final owner = tester //
+        .state<ProviderScopeState>(find.byType(ProviderScope))
+        .owner;
+
+    provider0.state.readOwner(owner);
+    provider1.state.readOwner(owner);
+    final familyState0 = owner.debugProviderStates.firstWhere((p) {
+      return p.provider == provider0.state;
+    });
+    final familyState1 = owner.debugProviderStates.firstWhere((p) {
+      return p.provider == provider1.state;
+    });
+
+    expect(buildCount, 1);
+    expect(familyState0.$hasListeners, true);
+    expect(familyState1.$hasListeners, false);
+    expect(find.text('0'), findsOneWidget);
+
+    notifier0.increment();
+    await tester.pump();
+
+    expect(buildCount, 2);
+    expect(find.text('1'), findsOneWidget);
+
+    notifier1.increment();
+    await tester.pump();
+
+    expect(buildCount, 2);
+
+    // changing the provider that computed is subscribed to
+    stateProvider.readOwner(owner).state = 1;
+    await tester.pump();
+
+    expect(buildCount, 3);
+    expect(find.text('43'), findsOneWidget);
+    expect(familyState1.$hasListeners, true);
+    expect(familyState0.$hasListeners, false);
+
+    notifier1.increment();
+    await tester.pump();
+
+    expect(buildCount, 4);
+    expect(find.text('44'), findsOneWidget);
+
+    notifier0.increment();
+    await tester.pump();
+
+    expect(buildCount, 4);
+  });
+  testWidgets('Consumer supports changing the provider', (tester) async {
+    final notifier1 = TestNotifier();
+    final provider1 = StateNotifierProvider((_) => notifier1);
+    final notifier2 = TestNotifier(42);
+    final provider2 = StateNotifierProvider((_) => notifier2);
+    var buildCount = 0;
+
+    Widget build(StateNotifierProvider<TestNotifier> provider) {
+      return ProviderScope(
+        child: Consumer((c, read) {
+          buildCount++;
+          final value = read(provider.state);
+          return Text('$value', textDirection: TextDirection.ltr);
+        }),
+      );
+    }
+
+    await tester.pumpWidget(build(provider1));
+
+    expect(find.text('0'), findsOneWidget);
+    expect(buildCount, 1);
+
+    await tester.pumpWidget(build(provider2));
+
+    expect(find.text('42'), findsOneWidget);
+    expect(buildCount, 2);
+
+    notifier1.increment();
+    await tester.pump();
+
+    expect(find.text('42'), findsOneWidget);
+    expect(buildCount, 2);
+
+    notifier2.increment();
+    await tester.pump();
+
+    expect(find.text('43'), findsOneWidget);
+    expect(buildCount, 3);
+  });
   testWidgets(
       'mutliple useProviders/read, when one of them forces rebuild, all dependencies are still flushed',
       (tester) async {
@@ -347,4 +531,8 @@ class TestNotifier extends StateNotifier<int> {
 
   // ignore: avoid_setters_without_getters
   set value(int value) => state = value;
+}
+
+class Listener<T> extends Mock {
+  void call(T value);
 }