+++ packages/riverpod/lib/src/internals.dart
@@ -6,7 +6,6 @@
 library internals;
 
 export 'common.dart';
-export 'computed.dart';
 export 'framework/framework.dart';
 export 'future_provider/future_provider.dart';
 export 'provider/provider.dart';